consumer/requirements.txt
memphis_py==0.2.7
redis
pytest
mockito
asynctest
mockito